yeaphone: Optimized scheduling of periodic timers
If a periodic timer is created with optimization enabled, then its 
interval is compared to the interval of already scheduled periodic 
timers. If one of the intervals is similar then the new timer is 
scheduled so that timeouts occur at the exact same time as often as 
possible to reduce system wakeups.
This will be especially usefull when Rafal's code is merged (eg. for the 
blinking cursor).


git-svn-id: svn://devbase.homelinux.org:5070/voip/yeaphone/trunk@78 3bd21d39-fd32-0410-89e0-1de79b5a6eeb